+++ examples/toggle_button.rs
@@ -21,12 +21,12 @@ pub fn main() {
         .opengl()
         .build()
         .unwrap();
-    let mut engine = Engine::new();
+    let mut engine = Engine::new(400, 100);
     let mut button1 = ToggleButtonWidget::new(20, 20, 170, 60, String::from("1"), 40, false);
 
     button1.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button1.set_numeric(CONFIG_BORDER_WIDTH, 2);
-    button1.on_toggle(|x, _widgets, _state| {
+    button1.on_toggle(|x, _widgets, _layouts, _state| {
         eprintln!("1 Toggled: {}", _state);
     });
 
@@ -34,11 +34,9 @@ pub fn main() {
 
     button2.set_color(CONFIG_COLOR_BORDER, Color::RGB(0, 0, 0));
     button2.set_numeric(CONFIG_BORDER_WIDTH, 2);
-    button2.on_toggle(|x, _widgets, _state| {
+    button2.on_toggle(|x, _widgets, _layouts, _state| {
         eprintln!("2 Toggled: {}", _state);
     });
-
-    engine.setup(400, 100);
 
     engine.add_widget(Box::new(button1), String::from("button1"));
     engine.add_widget(Box::new(button2), String::from("button2"));