+++ src/render/engine.rs
@@ -14,17 +14,19 @@
 // limitations under the License.
 
 use sdl2::event::Event;
-//use sdl2::messagebox::*;
 use sdl2::video::Window;
 use sdl2::Sdl;
 
+use crate::render::layout::Layout;
+use crate::render::layout_cache::LayoutCache;
 use crate::render::widget::{BaseWidget, Widget};
 use crate::render::widget_cache::WidgetCache;
 use std::time::Duration;
 
 /// This is a storage container for the Pushrod event engine.
 pub struct Engine {
-    cache: WidgetCache,
+    widget_cache: WidgetCache,
+    layout_cache: LayoutCache,
     current_widget_id: i32,
 }
 
@@ -51,26 +53,32 @@ pub struct Engine {
 /// That's all there is to it.  If you want to see more interactions on how the `Engine` is used in
 /// an application, check out the demo test code, and look at `rust-pushrod-chassis`.
 impl Engine {
-    /// Creates a new `Engine` object.
-    pub fn new() -> Self {
+    /// Creates a new `Engine` object.  Sets the engine up with the bounds of the screen, which
+    /// must be provided at instantiation time.  This is in order to set up the `BaseWidget` in the
+    /// top-level of the `Engine`, so that it knows what area of the screen to refresh when
+    /// required as part of the draw cycle.
+    pub fn new(w: u32, h: u32) -> Self {
+        let base_widget = BaseWidget::new(0, 0, w, h);
+        let mut cache = WidgetCache::default();
+
+        cache.add_widget(Box::new(base_widget), "base".to_string());
+
         Self {
-            cache: WidgetCache::new(),
+            widget_cache: cache,
+            layout_cache: LayoutCache::default(),
             current_widget_id: 0,
         }
     }
 
-    /// Sets up the top-level widget so that other widgets can be added to the screen.
-    pub fn setup(&mut self, window_width: u32, window_height: u32) {
-        let base_widget = BaseWidget::new(0, 0, window_width, window_height);
-
-        self.cache
-            .add_widget(Box::new(base_widget), "base".to_string());
-    }
-
-    /// Adds a widget to the display list.  Widgets are rendered in the order in which they were
+    /// Adds a `Widget` to the display list.  `Widget`s are rendered in the order in which they were
     /// created in the display list.
     pub fn add_widget(&mut self, widget: Box<dyn Widget>, widget_name: String) -> i32 {
-        self.cache.add_widget(widget, widget_name)
+        self.widget_cache.add_widget(widget, widget_name)
+    }
+
+    /// Adds a `Layout` to the `Layout` list.
+    pub fn add_layout(&mut self, layout: Box<dyn Layout>) -> i32 {
+        self.layout_cache.add_layout(layout)
     }
 
     /// Main application run loop, controls interaction between the user and the application.
@@ -88,87 +96,76 @@ impl Engine {
                     Event::MouseButtonDown {
                         mouse_btn, clicks, ..
                     } => {
-                        self.cache.button_clicked(
+                        self.widget_cache.button_clicked(
                             self.current_widget_id,
                             mouse_btn as u8,
                             clicks,
                             true,
+                            self.layout_cache.get_layout_cache(),
                         );
                     }
 
                     Event::MouseButtonUp {
                         mouse_btn, clicks, ..
                     } => {
-                        self.cache
-                            .button_clicked(-1, mouse_btn as u8, clicks, false);
+                        self.widget_cache.button_clicked(
+                            -1,
+                            mouse_btn as u8,
+                            clicks,
+                            false,
+                            self.layout_cache.get_layout_cache(),
+                        );
                     }
 
                     Event::MouseMotion { x, y, .. } => {
                         let cur_widget_id = self.current_widget_id;
 
-                        self.current_widget_id = self.cache.find_widget(x, y);
+                        self.current_widget_id = self.widget_cache.find_widget(x, y);
 
                         if cur_widget_id != self.current_widget_id {
-                            self.cache.mouse_exited(cur_widget_id);
-                            self.cache.mouse_entered(self.current_widget_id);
+                            self.widget_cache
+                                .mouse_exited(cur_widget_id, self.layout_cache.get_layout_cache());
+                            self.widget_cache.mouse_entered(
+                                self.current_widget_id,
+                                self.layout_cache.get_layout_cache(),
+                            );
                         }
 
-                        self.cache.mouse_moved(self.current_widget_id, vec![x, y]);
+                        self.widget_cache.mouse_moved(
+                            self.current_widget_id,
+                            vec![x, y],
+                            self.layout_cache.get_layout_cache(),
+                        );
                     }
 
                     Event::MouseWheel { x, y, .. } => {
-                        self.cache
-                            .mouse_scrolled(self.current_widget_id, vec![x, y]);
+                        self.widget_cache.mouse_scrolled(
+                            self.current_widget_id,
+                            vec![x, y],
+                            self.layout_cache.get_layout_cache(),
+                        );
                     }
 
                     Event::Quit { .. } => {
-                        //                        let buttons: Vec<_> = vec![
-                        //                            ButtonData {
-                        //                                flags: MessageBoxButtonFlag::RETURNKEY_DEFAULT,
-                        //                                button_id: 1,
-                        //                                text: "Yes",
-                        //                            },
-                        //                            ButtonData {
-                        //                                flags: MessageBoxButtonFlag::ESCAPEKEY_DEFAULT,
-                        //                                button_id: 2,
-                        //                                text: "No",
-                        //                            },
-                        //                        ];
-                        //
-                        //                        let res = show_message_box(
-                        //                            MessageBoxFlag::WARNING,
-                        //                            buttons.as_slice(),
-                        //                            "Quit",
-                        //                            "Are you sure?",
-                        //                            None,
-                        //                            None,
-                        //                        )
-                        //                        .unwrap();
-                        //
-                        //                        if let ClickedButton::CustomButton(x) = res {
-                        //                            if x.button_id == 1 {
                         break 'running;
-                        //                            }
-                        //                        }
                     }
 
                     remaining_event => {
-                        self.cache
-                            .other_event(self.current_widget_id, remaining_event);
+                        self.widget_cache.other_event(
+                            self.current_widget_id,
+                            remaining_event,
+                            self.layout_cache.get_layout_cache(),
+                        );
                     }
                 }
             }
 
-            self.cache.tick();
-            self.cache.draw_loop(&mut canvas);
+            self.widget_cache.tick(self.layout_cache.get_layout_cache());
+            self.layout_cache
+                .do_layout(self.widget_cache.borrow_cache());
+            self.widget_cache.draw_loop(&mut canvas);
 
             ::std::thread::sleep(Duration::new(0, 1_000_000_000u32 / 60));
         }
-    }
-}
-
-impl Default for Engine {
-    fn default() -> Self {
-        Self::new()
     }
 }