+++ src/widgets/image_widget.rs
@@ -14,6 +14,7 @@
 // limitations under the License.
 
 use crate::render::callbacks::CallbackRegistry;
+use crate::render::layout_cache::LayoutContainer;
 use crate::render::widget::*;
 use crate::render::widget_cache::WidgetContainer;
 use crate::render::widget_config::{